feat: rebrand to GameDesk — rename package, SO library, and app icons
- Rename Android package from com.carriez.flutter_hbb to me.blocker.gamedesk
  (Kotlin sources, AndroidManifest.xml, ffi.kt import)
- Rename SO library from librustdesk to libgamedesk across Cargo.toml,
  Rust sources, Flutter FFI, build scripts, and Xcode projects
- Replace Android launcher icons with new app icon at all DPI densities
