+++ src/plugin/native.rs
@@ -10,10 +10,10 @@ use super::{
     errno::ERR_NOT_HANDLED,
     native_handlers::{Callable, NATIVE_HANDLERS_REGISTRAR},
 };
-/// The native returned value from librustdesk native.
+/// The native returned value from libgamedesk native.
 ///
 /// [Note]
-/// The data is owned by librustdesk.
+/// The data is owned by libgamedesk.
 #[repr(C)]
 pub struct NativeReturnValue {
     pub return_type: c_int,
